Update weibo preprocess pipeline
