+++ data/MM17-WeiboRumorSet.dvc
@@ -1,5 +1,5 @@
 outs:
-- md5: d61ac8a5261e3d11b7573453439c0f15.dir
-  size: 2264658414
-  nfiles: 26546
+- md5: a3f3b30e68a94b68eb695cc40fd96a11.dir
+  size: 2264532157
+  nfiles: 26545
   path: MM17-WeiboRumorSet
